+++ app/src/pitstop/java/com/pitstop/models/TripIndicator.java
@@ -10,10 +10,10 @@ import com.pitstop.network.RequestCallback;
 
 public abstract class TripIndicator {
 
-    int tripId;
+    long tripId;
     String rtcTime;
 
-    TripIndicator(int tripId, String rtcTime) {
+    TripIndicator(long tripId, String rtcTime) {
         this.tripId = tripId;
         this.rtcTime = rtcTime;
     }
